+++ conftest.py
@@ -6,6 +6,7 @@ import pandas as pd
 import pytest
 
 import eniric
+from eniric.atmosphere import Atmosphere
 from eniric.IOmodule import pdread_2col
 
 resampled_template = "Spectrum_{0}-PHOENIX-ACES_{1}band_vsini{2}_R{3}_res3.0.txt"
@@ -114,3 +115,37 @@ def trans_unit(request):
 def grad_flag(request):
     """Gradient flag parameter."""
     return request.param
+
+
+@pytest.fixture(
+    params=[
+        # "Average_TAPAS_2014_H.txt",
+        "Average_TAPAS_2014_K.txt",
+        "Average_TAPAS_2014_J.txt",
+    ]
+)
+def atm_model(request):
+    """Get atmospheric model name to load."""
+    return os.path.join(eniric.paths["atmmodel"], request.param)
+
+
+@pytest.fixture(params=[2.5, 4])
+def atmosphere_fixture(request, atm_model):
+    percent_cutoff = request.param
+    atm = Atmosphere.from_file(atm_model)
+    atm.mask_transmission(percent_cutoff)
+    return atm
+
+
+@pytest.fixture()
+def short_atmosphere(atmosphere_fixture):
+    # First 2000 data points only to speed up tests
+    return atmosphere_fixture[:2000]
+
+
+@pytest.fixture(params=[(0, 1500), (8000, 9000)])
+def sliced_atmmodel_default_mask(request, atm_model):
+    """To do own masking. Sliced in different places."""
+    lower, upper = request.param  # slice limits
+    atm = Atmosphere.from_file(atm_model)
+    return atm[int(lower) : int(upper)]
